[refactor/#31] roomId 타입 통일

+++ core/data/chatting/src/commonMain/kotlin/com/nexters/emotia/core/data/chatting/repository/ChattingRepositoryImpl.kt
@@ -27,7 +27,7 @@ class ChattingRepositoryImpl(
 
     override suspend fun sendChat(roomId: String, message: String): Result<ChatMessage> {
         return runCatching {
-            val response = remoteDataSource.sendChat(roomId.toInt(), message)
+            val response = remoteDataSource.sendChat(roomId, message)
 
             if (!response.success) {
                 throw IllegalStateException("API 요청 실패: ${response.error}")
